2.1.0 (#64)
<!-- Korbit AI PR Description Start -->
## Description by Korbit AI

### What change is being made?

Update the `cordova-plugin-applovin-max` to version 2.1.0, upgrading
dependencies to Android SDK v13.2.0 and iOS SDK v13.2.0, and reflecting
these changes in the plugin's metadata files.

### Why are these changes being made?

These changes are made to incorporate enhancements and fixes provided in
the new versions of the Android and iOS SDKs, thereby ensuring that the
plugin is up-to-date with the latest functionality and improvements.
Updating the version numbers in the metadata and changelog ensures
accurate documentation and version tracking.

> Is this description stale? Ask me to generate a new description by
commenting `/korbit-generate-pr-description`
<!-- Korbit AI PR Description End -->